Start building skybox

+++ internal/lib/tiledStructs.go
@@ -1,10 +1,10 @@
 package lib
 
 type MapProps struct {
-	Bg        string `json:"bg"`
-	Horizon   int    `json:"horizon"`
-	NextLevel int    `json:"next level"`
-	TopLimit  int    `json:"top limit"`
+	Bg        string  `json:"bg"`
+	Horizon   float32 `json:"horizon"`
+	NextLevel int     `json:"next level"`
+	TopLimit  int     `json:"top limit"`
 }
 
 // Stores layer data: Data (a tile layer)
